LPS-35898 Move NULL_ACCESS_CONTROLLED from AccessControlled to AccessControl
