+++ portal-service/src/com/liferay/portal/security/ac/AccessControl.java
@@ -19,6 +19,8 @@ import com.liferay.portal.kernel.exception.SystemException;
 import com.liferay.portal.security.auth.AuthException;
 import com.liferay.portal.security.auth.AuthVerifierResult;
 
+import java.lang.annotation.Annotation;
+
 import java.util.Map;
 
 import javax.servlet.http.HttpServletRequest;
@@ -37,5 +39,24 @@ public interface AccessControl {
 
 	public AuthVerifierResult.State verifyRequest()
 		throws PortalException, SystemException;
+
+	public AccessControlled NULL_ACCESS_CONTROLLED = new AccessControlled() {
+
+		@Override
+		public Class<? extends Annotation> annotationType() {
+			return AccessControlled.class;
+		}
+
+		@Override
+		public boolean guestAccessEnabled() {
+			return false;
+		}
+
+		@Override
+		public boolean hostAllowedValidationEnabled() {
+			return false;
+		}
+
+	};
 
 }
